pocs/requirements.txt
pyftdi
pyserial
